requirements.txt
python-telegram-bot==13.8
pillow==8.4
